+++ src/ladb_opencutlist/ruby/constants.rb
@@ -2,6 +2,6 @@ module Ladb::OpenCutList
 
     EXTENSION_NAME = 'OpenCutList'.freeze
     EXTENSION_VERSION = '1.6.2-dev'.freeze
-    EXTENSION_BUILD = '201903031641'.freeze
+    EXTENSION_BUILD = '201904231315'.freeze
 
 end