+++ CHANGELOG.md
@@ -8,6 +8,7 @@ in 1.x and 0.x versions.
 
   * Add raw volume in all type of material in cutlist summary
   * Add cutlist "width" separators if first sort order is width on solid wood material parts
+  * Fixed Highlight tool bug with simple ComponentInstance children
   
 * 1.6.1 (2019-02-15)